art_guessing/frontend/requirements.txt
streamlit
PIL
matplotlib.pyplot
